Implemented resuable command object.
A command object can be constructed, as follows:

    command test
        pre-up
            echo up %1%
        post-down
            echo down %1%

The command can then be used from within an interface or group stanza.

    iface myiface
        command test something
        command test something2

The words following the command will be expanded and the inserted into
the command lines for the stanza.  up lines are added at the bottom,
while down lines are added at the top.