Changed how lines are parsed.
Originally, a stanza such as:

    define something
        pre-up
            echo 2
        pre-down
            echo 2

    iface blah
        pre-up
            echo 1
        pre-down
            echo 1
        call something
        pre-up
            echo 3
        pre-down
            echo 3

Would echo 1, 2, 3 during the up phase, but echo 2 1 3 during the down
phase.  THe parser has been changed so lines from down sections are
inserted at the top of the command lines (but still in the order for
that section.  So now, the down command will echo 3, 2, 1.