Implemented mappings
Mappings are now implemented.  The following syntax is used:

    mapping <pattern> [<pattern> ...]
        script /path/to/script
        map LINE
        ..

The script is executed with the interface/group name as the argument.
Each map line is sent to the standard input (without the "map" portion.
The output of the script is used as the logical definition from the file
to use.