Shut down interface and handle groups correctly.
When shutting down, all items in the state file are put down in reverse
order.  However, it is possible that an item in the state file may be a
group and put down other items in the state file, which will cause an
error when trying to put them down twice.  So instead, each iteration
checks the last item in the state file to prevent this.